20241123 나의 작업내용과 팀원의 작업내용을 잠정적으로 통합함
