+++ Sns/src/main/webapp/WEB-INF/views/main/viewcomment.jsp
